requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
joblib
shap
streamlit
plotly
mysql-connector-python
seaborn
matplotlib